Create new BookForm component
- Refactor BooksPage component
