+++ src/components/BooksPage.js
@@ -1,23 +1,8 @@
-import { useSelector, useDispatch } from 'react-redux';
-import { v4 as uuidv4 } from 'uuid';
+import { useSelector } from 'react-redux';
 import Book from './Book';
-import { addBook } from '../redux/books/books';
+import BookForm from './BookForm';
 
 const BooksPage = () => {
-  const dispatch = useDispatch();
-  const submitForm = (e) => {
-    e.preventDefault();
-    const title = e.target.querySelector('#input-title');
-    const author = e.target.querySelector('#input-author');
-    const data = {
-      id: uuidv4(),
-      title: title.value,
-      author: author.value,
-    };
-    title.value = '';
-    author.value = '';
-    dispatch(addBook(data));
-  };
   const bookList = useSelector((state) => state.books);
 
   return (
@@ -26,11 +11,7 @@ const BooksPage = () => {
       <ul>
         {bookList.map((book) => <Book key={book.id} book={book} />)}
       </ul>
-      <form id="add-book-form" onSubmit={(e) => submitForm(e)}>
-        <input id="input-title" placeholder="Book Title" />
-        <input id="input-author" placeholder="Book Author" />
-        <button type="submit"> Add Book </button>
-      </form>
+      <BookForm />
     </div>
   );
 };
